Replaced `makeheaders` and `xxd` tools with direct embeding using #embed, clang v20 required
Added warnings to ignore for dependencies, caused by clang v20.

+++ src/ui/assets.hpp
@@ -1,2 +1,3 @@
+
 extern const unsigned char splash_image[];
-extern const long long splash_size;
+extern const size_t splash_size;
